requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asttokens==2.4.1
attrs==24.2.0
av==12.3.0
Brotli
certifi
cffi==1.17.1
charset-normalizer
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.4.0
datasets==3.0.1
decorator==5.1.1
dill==0.3.8
executing==2.1.0
fastapi==0.115.0
faster-whisper==1.0.3
filelock
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gmpy2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.7
iniconfig==2.0.0
ipython==8.27.0
jedi==0.19.1
Jinja2
MarkupSafe
matplotlib-inline==0.1.7
mkl-service
mkl_fft
mkl_random
mpmath
multidict==6.1.0
multiprocess==0.70.16
networkx
numpy
onnxruntime==1.19.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow
pluggy==1.5.0
prompt_toolkit==3.0.48
protobuf==4.21.6
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyreadline3==3.5.4
PySocks
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.10
pytz==2024.2
PyYAML
regex==2024.9.11
requests
safetensors==0.4.5
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
stack-data==0.6.3
starlette==0.38.6
sympy
tokenizers==0.20.0
tqdm==4.66.5
traitlets==5.14.3
transformers
typing_extensions
tzdata==2024.2
urllib3
uvicorn==0.30.6
wcwidth==0.2.13
win-inet-pton
xxhash==3.5.0
yarl==1.13.0
intel-cmplr-lib-ur==2024.2.1
intel-openmp==2024.2.1
mkl==2024.2.2
tbb==2021.13.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
